Added mobile support and fixed resolution. Tested on both iOS and Android mobile devices

+++ space-invaders-7/src/Backup Files/Game.js
@@ -6,10 +6,15 @@ BasicGame.Game.prototype =
 {
 	create: function () 
 	{
+		this.playerSpeedRight = +150;
+		this.playerSpeedLeft = -150;
 		this.firingTimer = 0;
 		this.bulletTime = 0;
 		this.score = 0;
 		this.livingEnemies = [];
+		this.levelScore = 1;
+		this.laser = this.add.audio('laser');
+		this.explosionaudio = this.add.audio('explosionaudio');
 		this.physics.startSystem(Phaser.Physics.ARCADE);
 		this.starfield = this.add.tileSprite(0, 0, 800, 600, 'starfield');
 		this.bullets = this.add.group();
@@ -34,9 +39,16 @@ BasicGame.Game.prototype =
 		this.aliens = this.add.group();
 		this.aliens.enableBody = true;
 		this.aliens.physicsBodyType = Phaser.Physics.ARCADE;
+		this.level = false;
 		this.createAliens();
 		this.scoreString = 'Score : ';
-		this.scoreText = this.add.text(10, 10, this.scoreString + this.score, { font: '34px Arial', fill: '#fff' });
+		this.scoreText = this.add.text(this.game.width - 390, this.game.width - 790, this.scoreString + this.score, { font: '34px Arial', fill: '#fff' });
+		this.levelString = 'Level : ';
+		this.levelText = this.add.text(this.game.width - 800, this.game.width - 580, this.levelString + this.levelScore, { font: '34px Arial', fill: '#fff' });
+		this.health = 100;
+		this.healthString = 'Health : ';
+		this.healthText = this.add.text(400, 10, this.healthString + this.health, { font: '34px Arial', fill: '#fff' });
+
 		this.lives = this.add.group();
 		this.add.text(this.world.width - 100, 10, 'Lives : ', { font: '34px Arial', fill: '#fff' });
 		this.stateText = this.add.text(this.world.centerX,this.world.centerY,' ', { font: '84px Arial', fill: '#fff' });
@@ -54,60 +66,84 @@ BasicGame.Game.prototype =
 		this.explosions.createMultiple(30, 'kaboom');
 		this.explosions.forEach(this.setupInvader, this);
 		this.cursors = this.input.keyboard.createCursorKeys();
-		this.fireButton = this.input.keyboard.addKey(Phaser.Keyboard.SPACEBAR);
 		this.scale.fullScreenScaleMode = Phaser.ScaleManager.EXACT_FIT;
-
-		var fullscreen = this.add.button(this.game.width-8, this.game.height-8,'fullscreen',BasicGame.toggleFullscreen,this,'over', 'up', 'down');
+		var fullscreen = this.add.button(this.game.width-25, this.game.height-25,'fullscreen',BasicGame.toggleFullscreen,this,'over', 'up', 'down');
 		fullscreen.pivot.x = fullscreen.width;
 		fullscreen.pivot.y = fullscreen.height;
+		this.scale.pageAlignHorizontally = true;
+        this.scale.pageAlignVertically = true;
+		this.scale.refresh();
+   		this.world.setBounds(0, 0, 800, 600);
+		this.player.body.collideWorldBounds = true;
 
-		this.leftButton = this.add.button(this.world.centerX - 200, this.world.centerY + 290,'left', null, this, 2, 1);
-		this.leftButton.pivot.x = this.leftButton.width;
-		this.leftButton.pivot.y = this.leftButton.height;
+		this.healthbar = this.game.add.sprite(this.game.width - 325, this.game.height - 540 ,'healthbar');
+		this.healthbar.cropEnabled = true;
+		this.healthbar.crop.width = (this.player.health / this.player.maxHealth) * this.healthbar.width;
 
-		this.rightButton = this.add.button(this.world.centerX + 200, this.world.centerY + 290,'right', null, this);
-		this.rightButton.pivot.x = this.rightButton.width;
-		this.rightButton.pivot.y = this.rightButton.height;
+		this.healthbar.animations.add('healthbar');
+		this.healthbar.animations.play('healthbar', 30, false); 
 
-		this.fireClick = this.add.button(this.world.centerX, this.world.centerY + 290, 'fireClick', null, this, 2, 1);
-		this.fireClick.pivot.x = this.fireClick.width;
-		this.fireClick.pivot.y = this.fireClick.height;
+		
 
 	},
 
 	update: function () 
 	{
 		this.starfield.tilePosition.y += 2;
+		var xInput = this.input.x;
+		this.fireBullet();
+
 		if(this.player.alive)
 		{
-			this.player.body.velocity.setTo(0, 0);
-		  
-			if (this.cursors.left.isDown || this.leftButton.input.pointerDown(1) || this.leftButton.input.pointerDown())
+			if (this.game.device.desktop) 
 			{
-				this.player.body.velocity.x = -200;
-			}
-			else if (this.cursors.right.isDown || this.rightButton.input.pointerDown(1) || this.rightButton.input.pointerDown())
+				if (this.cursors.right.isDown) 
+				{
+					this.player.body.velocity.x = this.playerSpeedRight;
+				}
+
+				if (this.cursors.left.isDown) 
+				{
+					this.player.body.velocity.x = this.playerSpeedLeft;
+				}
+			} 	
+			else
 			{
-				this.player.body.velocity.x = 200;
+				if (xInput > 400) 
+				{
+					this.player.body.velocity.x = this.playerSpeedRight;
+				}
+
+				if (xInput <= 400) 
+				{
+					this.player.body.velocity.x = this.playerSpeedLeft;
+				}
 			}
-			if (this.fireButton.isDown || this.fireClick.input.pointerDown(1) || this.fireClick.input.pointerDown())
-			{
-				this.fireBullet();	
-			}
+
 			if (this.time.now > this.firingTimer)
 			{
-				this.enemyFires();
+				for (var i = 0; i < this.levelScore; i++)
+				{
+					this.enemyFires();
+				}
 			}
+
+			if (this.aliens.y > 600)
+			{
+				this.player.kill();
+				this.enemyBullets.callAll('kill'); //
+				this.stateText.text=" GAME OVER, \n Click to restart, \n Nice try human";
+				this.levelScore = 1;
+				this.stateText.visible = true;
+				this.input.onTap.addOnce(this.restartState,this);
+			}
+
+
+
 			this.physics.arcade.overlap(this.bullets, this.aliens, this.collisionHandler, null, this);
 			this.physics.arcade.overlap(this.enemyBullets, this.player, this.enemyHitsPlayer, null, this);
 			this.physics.arcade.overlap(this.player, this.aliens, this.playerCollision, null, this);
 		}
-	},
-
-
-	playerMoves: function()
-	{
-		this.player.body.velocity.x = -200;
 	},
 
 	restart: function () 
@@ -140,9 +176,11 @@ BasicGame.Game.prototype =
 		{
 			player.kill();
 			this.enemyBullets.callAll('kill'); //
-			this.stateText.text=" GAME OVER \n Click to restart";
+			this.stateText.text=" GAME OVER, \n Click to restart, \n Nice try human";
+			this.levelScore = 1;
 			this.stateText.visible = true;
-			this.input.onTap.addOnce(this.restart,this);
+			this.bullet.kill();
+			this.input.onTap.addOnce(this.restartState,this);
 		}
 	},
 
@@ -175,14 +213,22 @@ BasicGame.Game.prototype =
 		this.explosion = this.explosions.getFirstExists(false);
 		this.explosion.reset(alien.body.x, alien.body.y);
 		this.explosion.play('kaboom', 30, false, true);
+		this.explosionaudio.play();
+
 
 		if (this.aliens.countLiving() == 0)
 		{
 			this.score += 1000;
 			this.scoreText.text = this.scoreString + this.score;
-			this.stateText.text = " You Won, \n Click to restart";
+			this.stateText.text = " You Won, \n Click to move on";
 			this.stateText.visible = true;
 			this.enemyBullets.callAll('kill');
+			this.level = true;
+			this.levelScore = this.levelScore + 1;
+			this.levelText.text = this.levelString + this.levelScore;
+			this.playerSpeedRight += 25;
+			this.playerSpeedLeft -= 25;
+			bullet.kill();
 			this.input.onTap.addOnce(this.restart,this);
 		}
 	},
@@ -195,6 +241,7 @@ BasicGame.Game.prototype =
 
 			if (this.bullet)
 			{
+				this.laser.play();
 				this.bullet.reset(this.player.x, this.player.y + 8);
 				this.bullet.body.velocity.y = -400;
 				this.bulletTime = this.time.now + 200;
@@ -211,7 +258,18 @@ BasicGame.Game.prototype =
 		this.stateText.visible = false;
 	},
 
-	
+	gofull: function() 
+	{
+		if (this.scale.isFullScreen)
+		{
+			this.scale.stopFullScreen();
+		}
+
+		else
+		{
+			this.scale.startFullScreen(false);
+		}
+	},
 
 
 	resetBullet: function(bullet) 
@@ -233,6 +291,7 @@ BasicGame.Game.prototype =
 
 	createAliens: function () 
 	{
+
 		for (var y = 0; y < 4; y++)
 		{
 			for (var x = 0; x < 10; x++)
@@ -246,6 +305,11 @@ BasicGame.Game.prototype =
 		}
 		this.aliens.x = 100;
 		this.aliens.y = 50;
+
+		this.aliens.checkWorldBounds = true;
+		this.aliens.setAll('outOfBoundsKill', true);
+		this.aliens.setAll('checkWorldBounds', true);
+
 		this.tween = this.add.tween(this.aliens).to( { x: 200 }, 2000, Phaser.Easing.Linear.None, true, 0, 1000, true);
 		this.tween.onLoop.add(this.descend, this);
 	},
@@ -254,23 +318,63 @@ BasicGame.Game.prototype =
 	{
 		bullet.kill();
 		this.live = this.lives.getFirstAlive();
+		this.reduceHealth();
 
-		if (this.live)
-		{
-			this.live.kill();
-		}
 		var explosion = this.explosions.getFirstExists(false);
 		explosion.reset(player.body.x, player.body.y);
 		explosion.play('kaboom', 30, false, true);
 
-		if (this.lives.countLiving() < 1)
+		if (this.health <= 1)
 		{
-			player.kill();
-			this.enemyBullets.callAll('kill');
-			this.stateText.text=" GAME OVER \n Click to restart";
-			this.stateText.visible = true;
-			this.input.onTap.addOnce(this.restart,this);
+			this.health += 100;
+			if (this.live)
+			{
+				this.live.kill();
+			}
+			var explosion = this.explosions.getFirstExists(false);
+			explosion.reset(player.body.x, player.body.y);
+			explosion.play('kaboom', 30, false, true);
+			this.explosionaudio.play();
+			this.healthText.text = this.healthString + this.health;
+
+			if (this.lives.countLiving() < 1)
+			{
+				player.kill();
+				this.enemyBullets.callAll('kill');
+				this.stateText.text=" GAME OVER, \n Click to restart,\n Nice try human";
+				this.levelScore = 1;
+				this.stateText.visible = true;
+				this.input.onTap.addOnce(this.restartState,this);
+			}
 		}
+	},
+
+	reduceHealth: function()
+	{
+		this.health-=10;
+		this.healthText.text = this.healthString + this.health;
+
+		console.log(this.health);
+	},
+
+
+	AliensOutOfBounds: function() 
+	{
+		player.kill();
+		this.enemyBullets.callAll('kill');
+		this.stateText.text=" GAME OVER, \n Click to restart,\n Nice try human";
+		this.levelScore = 1;
+		this.stateText.visible = true;
+		this.input.onTap.addOnce(this.restartState,this);
+	},
+
+
+
+
+
+	restartState: function()
+	{
+		this.state.start(this.state.current);
 	},
 
 	quitGame: function (pointer)
